Palette Change
I wanted to change the palette so that it matched with the mood of my chosen song which has yet to be uploaded into my code.

However, I couldn't change the background to my wanted colour without disrupting the map sequence. I'll have to find a way to keep the map code but edit the colours.