"Randomising" Ripples + deleting the words
Using BASS as my trial, I wanted to experiment with randomising ripple placements before moving onto animating them. Using code randomiser, I realised that it looked quite erratic and didn't actually meet the brief of the visualiser reacting to the music. So instead I turned to if statements. So if bass was less than or equal to 25 a ripple would appear at the bottom left corner of the screen, and if it was greater than 25 it would appear at the center so on and so forth.